Fix JENKINS-47034
Report Name, directory, files, and titles are now all trimmed. Note that
trimming is only done when configuration is saved.
