+++ src/main/java/htmlpublisher/HtmlPublisherTarget.java
@@ -24,6 +24,7 @@ import javax.servlet.ServletException;
 
 import hudson.util.HttpResponses;
 import jenkins.model.RunAction2;
+import org.apache.commons.lang.StringUtils;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
 
@@ -111,10 +112,10 @@ public class HtmlPublisherTarget extends AbstractDescribableImpl<HtmlPublisherTa
      */
     @DataBoundConstructor
     public HtmlPublisherTarget(String reportName, String reportDir, String reportFiles,String reportTitles, boolean keepAll, boolean alwaysLinkToLastBuild, boolean allowMissing) {
-        this.reportName = reportName;
-        this.reportDir = reportDir;
-        this.reportFiles = reportFiles;
-        this.reportTitles = reportTitles;
+        this.reportName = StringUtils.trim(reportName);
+        this.reportDir = StringUtils.trim(reportDir);
+        this.reportFiles = StringUtils.trim(reportFiles);
+        this.reportTitles = StringUtils.trim(reportTitles);
         this.keepAll = keepAll;
         this.alwaysLinkToLastBuild = alwaysLinkToLastBuild;
         this.allowMissing = allowMissing;
@@ -132,12 +133,7 @@ public class HtmlPublisherTarget extends AbstractDescribableImpl<HtmlPublisherTa
      * @since 1.4
      */
     public HtmlPublisherTarget(String reportName, String reportDir, String reportFiles, boolean keepAll, boolean alwaysLinkToLastBuild, boolean allowMissing) {
-        this.reportName = reportName;
-        this.reportDir = reportDir;
-        this.reportFiles = reportFiles;
-        this.keepAll = keepAll;
-        this.alwaysLinkToLastBuild = alwaysLinkToLastBuild;
-        this.allowMissing = allowMissing;
+        this(reportName, reportDir, reportFiles, null, keepAll, alwaysLinkToLastBuild, allowMissing);
     }
 
     public String getReportName() {
