+++ src/test/groovy/htmlpublisher/HtmlPublisherIntegrationTest.groovy
@@ -12,7 +12,7 @@ import org.jvnet.hudson.test.TestBuilder
  *
  * @author Kohsuke Kawaguchi
  */
-public class HtmlPublisherTest extends HudsonTestCase {
+public class HtmlPublisherIntegrationTest extends HudsonTestCase {
     /**
      * Makes sure that the configuration survives the round trip.
      */
@@ -66,14 +66,5 @@ public class HtmlPublisherTest extends HudsonTestCase {
         assertFalse(tab2Files.contains("dummy.html"))
     }
 
-    public void testDefaultIncludes() {
-        HtmlPublisherTarget target1 = new HtmlPublisherTarget("tab1", "target", "tab1.html", true, true, false);
-        assertEquals(HtmlPublisherTarget.INCLUDE_ALL_PATTERN, target1.getIncludes());
-        target1.setIncludes(null);
-        assertEquals(HtmlPublisherTarget.INCLUDE_ALL_PATTERN, target1.getIncludes());
-        target1.setIncludes("hello");
-        assertEquals("hello", target1.getIncludes());
-        target1.setIncludes("");
-        assertEquals(HtmlPublisherTarget.INCLUDE_ALL_PATTERN, target1.getIncludes());
-    }
+
 }
